Replace the variable type indicators with explode modifiers that can behave according to variable type or externally defined schema.
Rationale: the comments received so far on the pre-draft
suggest that variable types add too much complexity in terms
of name processing yet not enough power for folks working
with more elaborate schemas. It is therefore simpler to just
include a couple value modifiers that indicate something
magic is about to happen and provide a default set of
magic for the common variable types.

-nd -This line, and those below, will be ignored--

M    draft-gregorio-uritemplate.xml